refactor: Controller 분리, Student entity 메서드 추가
- UserController에서 Profile 관련 기능 ProfileController로 이동
- Student에 getCourse메서드 추가
